Various fixes related to restructuring of MLT impl
- remove duplicated classes in shadowed jars
- move ThreadStackAccess related classes to mlt-agent (solves class loading issues)
- fix MLT initialization in Agent
