+++ dd-java-agent/agent-mlt/src/main/java/com/datadog/mlt/sampler/JmxThreadStackProvider.java
@@ -1,10 +1,10 @@
-package datadog.trace.core.util;
+package com.datadog.mlt.sampler;
 
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadInfo;
 import java.lang.management.ThreadMXBean;
 
-public class JmxThreadStackProvider implements ThreadStackProvider {
+final class JmxThreadStackProvider implements ThreadStackProvider {
   private final ThreadMXBean threadMXBean = ManagementFactory.getThreadMXBean();
 
   public static final ThreadStackProvider INSTANCE = new JmxThreadStackProvider();
@@ -14,7 +14,7 @@ public class JmxThreadStackProvider implements ThreadStackProvider {
     // TODO maxDepth should be configurable
     ThreadInfo[] threadInfos = threadMXBean.getThreadInfo(threadIds, 128);
     if (threadInfos.length == 0) {
-      return ThreadStackProvider.EMPTY_STACKTRACE_ARRAY;
+      return EMPTY_STACKTRACE_ARRAY;
     }
     StackTraceElement[][] stackTraces = new StackTraceElement[threadInfos.length][];
     for (int i = 0; i < threadInfos.length; i++) {
@@ -28,7 +28,7 @@ public class JmxThreadStackProvider implements ThreadStackProvider {
     // TODO maxDepth should be configurable
     ThreadInfo[] threadInfos = threadMXBean.getThreadInfo(threadIds, 128);
     if (threadInfos.length == 0) {
-      return ThreadStackProvider.EMPTY_THERADINFO_ARRAY;
+      return EMPTY_THERADINFO_ARRAY;
     }
     return threadInfos;
   }
