+++ dd-java-agent/agent-mlt/src/main/java/com/datadog/mlt/sampler/ThreadStackProvider.java
@@ -1,4 +1,4 @@
-package datadog.trace.core.util;
+package com.datadog.mlt.sampler;
 
 import java.lang.management.ThreadInfo;
