+++ dd-java-agent/agent-mlt/src/test/java/com/datadog/mlt/sampler/JMXSessionTest.java
@@ -2,7 +2,6 @@ package com.datadog.mlt.sampler;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import datadog.trace.core.util.ThreadStackAccess;
 import datadog.trace.mlt.Session;
 import org.junit.jupiter.api.Test;
