+++ dd-java-agent/agent-profiling/src/main/java/com/datadog/profiling/agent/ProfilingAgent.java
@@ -7,7 +7,6 @@ import com.datadog.profiling.controller.ProfilingSystem;
 import com.datadog.profiling.controller.UnsupportedEnvironmentException;
 import com.datadog.profiling.uploader.ProfileUploader;
 import datadog.trace.api.Config;
-import datadog.trace.mlt.MethodLevelTracer;
 import java.lang.ref.WeakReference;
 import java.time.Duration;
 import java.util.function.Predicate;
@@ -115,8 +114,6 @@ public class ProfilingAgent {
       if (uploader != null) {
         uploader.shutdown();
       }
-
-      MethodLevelTracer.shutdown();
     }
   }
 }
