--- dd-trace-core/src/main/java/datadog/trace/core/util/NoneThreadStackProvider.java
@@ -1,16 +1,0 @@
-package datadog.trace.core.util;
-
-import java.lang.management.ThreadInfo;
-
-public class NoneThreadStackProvider implements ThreadStackProvider {
-
-  @Override
-  public StackTraceElement[][] getStackTrace(long[] threadIds) {
-    return ThreadStackProvider.EMPTY_STACKTRACE_ARRAY;
-  }
-
-  @Override
-  public ThreadInfo[] getThreadInfo(long[] threadIds) {
-    return ThreadStackProvider.EMPTY_THERADINFO_ARRAY;
-  }
-}
