+++ utils/mlt-support/mlt-support.gradle
@@ -14,15 +14,20 @@ dependencies {
   testCompile group: 'nl.jqno.equalsverifier', name: 'equalsverifier', version: '3.3'
 }
 
+tasks.withType(Test).configureEach {
+  useJUnitPlatform()
+}
+
+// TODO temporarily lowering the coverage limits - the support classes are additionally tested in mlt-agent project
+minimumInstructionCoverage = 0.5
+minimumBranchCoverage = 0.5
+
+excludedClassesCoverage += [
+  // MLTReader and MLTWriter are exercised in mlt-agent project tests
+  'com.datadog.mlt.io.MLTWriter',
+  'com.datadog.mlt.io.MLTReader',
+  'com.datadog.mlt.io.MLTConstants'
+]
+
 sourceCompatibility = JavaVersion.VERSION_1_8
 targetCompatibility = JavaVersion.VERSION_1_8
-
-shadowJar {
-  configurations = [project.configurations.compile]
-  archiveClassifier.set('')
-
-  dependencies {
-    include(dependency('it.unimi.dsi:fastutil:8.2.1'))
-  }
-  minimize()
-}
